保留本分支安卓防盗版陷阱拆除与 nativeOnPause 守卫。
ActionManager::setClone 冲突取 main 的 intptr_t 转型与 _monsterArray 空守卫。

+++ projects/NarutoSenki/Classes/MyUtils/CMD5Checksum.h
@@ -12,10 +12,10 @@
       
       
      *****************************************************************************************/  
-    typedef unsigned long DWORD;  
+    typedef unsigned int DWORD;  
     typedef unsigned char BYTE;  
-    typedef unsigned long ULONG;  
-    typedef unsigned long ULONG;  
+    typedef unsigned int ULONG;  
+    typedef unsigned int ULONG;  
     typedef unsigned int UINT;  
     typedef unsigned char UCHAR;
